Use exported BodyPart enum for the buttons
That way, you have a single script for all body part button.

Though, I think that if you want to keep it super simpler, you might
want to instead have the monster body_part assignable directly for now.
As I'm not sure if the monster maker should need a middle-man between it
and the body part it is changing.

Anyway, have a great day :D